[Finish #120766683] UI enhancements
Make the update/create document dialog scrollable.
Make the main app bar fixed at the top.
Add a z-depth to documents
Change the title in the app bars to the name of the application.
Restructure the login page and make it responsive.
Link the user profile to the menu item next to the logout button.